chore: clean up components+names
